+++ src/GUI/PopupBase.hpp
@@ -4,13 +4,13 @@
 
 enum class MenuAnimation
 {
-    None,
-    FromTop,
-    FromBottom,
-    FromLeft,
-    FromRight,
-    Scale,
-    FadeIn
+    None = 0,
+    FromTop = 1,
+    FromBottom = 2,
+    FromLeft = 3,
+    FromRight = 4,
+    Scale = 5,
+    FadeIn = 6
 };
 
 class PopupBase : public cocos2d::CCLayerColor, public cocos2d::CCActionTweenDelegate
@@ -55,6 +55,7 @@ class PopupBase : public cocos2d::CCLayerColor, public cocos2d::CCActionTweenDel
         virtual bool setup();
         virtual void show();
         virtual void visit();
+        virtual void draw();
 
         void playAnimation(MenuAnimation anim);
         void onClose(CCObject* sender);